Fixes and updates to subdivision

+++ CogoTestBed/CogoTestBed/Node.cs
@@ -10,9 +10,9 @@ namespace CogoTestBed
     {
         #region Properties
 
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
 
         #endregion
